src/exchanges/okx/modules/authed/position/getLeverage.ts
import debug from 'debug'

import { IAlunaExchangeAuthed } from '../../../../../lib/core/IAlunaExchange'
import {
  IAlunaPositionGetLeverageParams,
  IAlunaPositionGetLeverageReturns,
} from '../../../../../lib/modules/authed/IAlunaPositionModule'
import { OkxHttp } from '../../../OkxHttp'
import { getOkxEndpoints } from '../../../okxSpecs'



const log = debug('alunajs:okx/position/getLeverage')



export const getLeverage = (exchange: IAlunaExchangeAuthed) => async (
  params: IAlunaPositionGetLeverageParams,
): Promise<IAlunaPositionGetLeverageReturns> => {

  const {
    settings,
    credentials,
  } = exchange

  const {
    id,
    symbolPair,
    http = new OkxHttp(settings),
  } = params

  log('getting leverage', { id, symbolPair })

  // TODO: Implement proper getter
  const leverage = await http.authedRequest<number>({
    credentials,
    url: getOkxEndpoints(settings).position.getLeverage,
    body: { id, symbolPair },
  })

  const { requestWeight } = http

  return {
    leverage,
    requestWeight,
  }

}
